Ajustando as validadações do delete e update

+++ src/exercicio/services/exercicio.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common'
 import { Exercicio } from '../entities/exercicio.entity'
 import { InjectRepository } from '@nestjs/typeorm'
 import { ILike, Repository } from 'typeorm'
@@ -8,59 +8,76 @@ import { DeleteResult } from 'typeorm'
 @Injectable()
 export class ExercicioService{
 
-   constructor(
+  constructor(
     @InjectRepository(Exercicio)
     private exercicioRepository: Repository<Exercicio>
-   ){}
+  ){}
 
-   async findAll(): Promise<Exercicio[]>{
+  async findAll(): Promise<Exercicio[]>{
     return await this.exercicioRepository.find()
 
-   }
-   async findById(id: number): Promise<Exercicio>{
-   const exercicio = await this.exercicioRepository.findOne({
+  }
+
+  async findById(id: number): Promise<Exercicio>{
+    const exercicio = await this.exercicioRepository.findOne({
     where: {id}})
+
     if (!exercicio){
-      throw new Error('Exercício não encontrado')
+      throw new HttpException("Exercício não encontrado", HttpStatus.NOT_FOUND)
     }
     return exercicio        
-   }
+  }
 
-   async findByGrupoMuscular(grupoMuscular : string) : Promise<Exercicio[]>{
+  async findByGrupoMuscular(grupoMuscular : string) : Promise<Exercicio[]>{
     return await this.exercicioRepository.find({
       where: {
         grupoMuscular: ILike(`%${grupoMuscular}%`)
       }
     })
-   }
+  }
 
-   async findByNomeExercicio(nomeExercicio : string) : Promise<Exercicio[]>{
+  async findByNomeExercicio(nomeExercicio : string) : Promise<Exercicio[]>{
     return await this.exercicioRepository.find({
       where: {
         nomeExercicio: ILike(`%${nomeExercicio}%`)
       }
     })
-   }
+  }
 
-   async findByNivel(nivel : string) : Promise<Exercicio[]>{
+  async findByNivel(nivel : string) : Promise<Exercicio[]>{
     return await this.exercicioRepository.find({
       where: {
         nivel: ILike(`%${nivel}%`)
       }
     })
-   }
+  }
 
-   async create(exercicio: Exercicio): Promise<Exercicio>{
+  async create(exercicio: Exercicio): Promise<Exercicio>{
+    
     return await this.exercicioRepository.save(exercicio)
-   }
+  }
 
-   async update(exercicio: Exercicio): Promise<Exercicio>{
+  async update(exercicio: Exercicio): Promise<Exercicio>{
+
+    const buscaExercicio: Exercicio = await this.findById(exercicio.id)
+
+    if(!buscaExercicio || !exercicio.id){
+      throw new HttpException("Exercício não encontrado", HttpStatus.NOT_FOUND)
+    }
+
     return await this.exercicioRepository.save(exercicio)
-   }
+  }
 
-   async delete(id: number): Promise<DeleteResult>{
+  async delete(id: number): Promise<DeleteResult>{
+
+    const buscaExercicio: Exercicio = await this.findById(id)
+
+    if(!buscaExercicio){
+      throw new HttpException("Exercício não encontrado", HttpStatus.NOT_FOUND)
+    }
+
     return await this.exercicioRepository.delete(id)
-   }
+  }
 
 }
